Rename to spi_rgb_led
