+++ esphome/components/spi_rgb_led/__init__.py
