+++ esphome/components/spi_rgb_led/light.py
@@ -4,12 +4,12 @@ from esphome.components import light
 from esphome.components import spi
 from esphome.const import CONF_OUTPUT_ID, CONF_NUM_LEDS, CONF_DATA_RATE
 
-apa102_ns = cg.esphome_ns.namespace("apa102")
-APA102 = apa102_ns.class_("APA102", light.AddressableLight, spi.SPIDevice)
+spi_rgb_led_ns = cg.esphome_ns.namespace("spi_rgb_led")
+SpiRgbLed = spi_rgb_led_ns.class_("SpiRgbLed", light.AddressableLight, spi.SPIDevice)
 
 CONFIG_SCHEMA = light.ADDRESSABLE_LIGHT_SCHEMA.extend(
     {
-        cv.GenerateID(CONF_OUTPUT_ID): cv.declare_id(APA102),
+        cv.GenerateID(CONF_OUTPUT_ID): cv.declare_id(SpiRgbLed),
         cv.Optional(CONF_NUM_LEDS, default=1): cv.positive_not_null_int,
         cv.Optional(CONF_DATA_RATE, default="1MHz"): spi.SPI_DATA_RATE_SCHEMA,
     }
@@ -23,4 +23,4 @@ async def to_code(config):
     await spi.register_spi_device(var, config)
     await cg.register_component(var, config)
     spi_data_rate = str(spi.SPI_DATA_RATE_OPTIONS[config[CONF_DATA_RATE]])
-    cg.add_define("APA102_DATA_RATE", cg.RawExpression(spi_data_rate))
+    cg.add_define("SPIRGBLED_DATA_RATE", cg.RawExpression(spi_data_rate))
